@@ -13,15 +13,16 @@
   ],
   "autoload": {
     "psr-4": {
-      "kouinkouin\StreetParser": "./"
+      "kouinkouin\\StreetParser\\": ""
     }
   },
   "require": {
-    "php": ">=5.3.0",
-    "ext-curl": "*"
+    "php": ">=5.4.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.*"
+    "phpunit/phpunit": "6.*"
+  },
+  "scripts": {
+    "test": "phpunit"
   }
-
 }
